Fix compilation warnings on Windows x64.

+++ include/tins/dot11/dot11_control.h
@@ -144,7 +144,9 @@ protected:
     /**
      * \brief Getter for the control ta additional fields size.
      */
-    uint32_t controlta_size() const { return _taddr.size() + sizeof(ieee80211_header); }
+    uint32_t controlta_size() const { 
+        return static_cast<uint32_t>(_taddr.size() + sizeof(ieee80211_header)); 
+    }
 
     uint32_t write_ext_header(uint8_t *buffer, uint32_t total_sz);
 private:
